+++ tcex/__init__.py
@@ -1,5 +1,5 @@
 __author__ = 'ThreatConnect (support@threatconnect.com)'
-__version__ = '0.7.17'
+__version__ = '0.7.18'
 __license__ = 'Apache License, Version 2'
 
 try:
@@ -9,4 +9,4 @@ except ImportError as e:
     print('Try running tc_lib')
 
 from .tcex_local import TcExLocal
-from .argparser import ArgParser
+from .tcex_argparser import TcExArgParser