+++ tcex/tcex_argparser.py
@@ -3,7 +3,7 @@
 from argparse import ArgumentParser
 
 
-class ArgParser(ArgumentParser):
+class TcExArgParser(ArgumentParser):
     """Overload of the ArgumentParser class.
 
     Adding common arguments for TcEx apps.
@@ -12,7 +12,7 @@ class ArgParser(ArgumentParser):
     def __init__(self, **kwargs):
         """Initialize default values for common args.
         """
-        super(ArgParser, self).__init__(**kwargs)
+        super(TcExArgParser, self).__init__(**kwargs)
 
         # api defaults
 
@@ -36,6 +36,7 @@ class ArgParser(ArgumentParser):
         self._tc_log_file = 'app.log'
         self._tc_log_path = '/tmp'
         self._tc_out_path = '/tmp'
+        self._tc_secure_params = 'false'
         self._tc_temp_path = '/tmp'
         self._tc_user_id = None
         self._tc_log_to_api = False
@@ -132,11 +133,14 @@ class ArgParser(ArgumentParser):
 
         --api_default_org org        The TC API user default organization.
         --tc_api_path path           The TC API path (e.g https://api.threatconnect.com).
-        --tc_out_path path           The app in path.
+        --tc_in_path path            The app in path.
+        --tc_log_file filename       The app log file name.
         --tc_log_path path           The app log path.
         --tc_out_path path           The app out path.
+        --tc_secure_params bool      Flag to indicator secure params is supported.
         --tc_temp_path path          The app temp path.
         --tc_user_id id              The user id of user running the job.
+
         --tc_proxy_host host         The proxy host.
         --tc_proxy_port port         The proxy port.
         --tc_proxy_username user     The proxy username.
@@ -161,6 +165,9 @@ class ArgParser(ArgumentParser):
             '--tc_log_path', default=self._tc_log_path, help='ThreatConnect log path')
         self.add_argument(
             '--tc_out_path', default=self._tc_out_path, help='ThreatConnect output path')
+        self.add_argument(
+            '--tc_secure_params', default=self._tc_secure_params,
+            help='ThreatConnect Secure params enabled')
         self.add_argument(
             '--tc_temp_path', default=self._tc_temp_path, help='ThreatConnect temp path')
         self.add_argument(