+++ tcex/tcex_job.py
@@ -453,7 +453,8 @@ class TcExJob(object):
                                     break
                                 else:
                                     # all indicators were saved minus failed; not_save == failed
-                                    self._indicator_results['not_saved'] = self._indicator_results.get('failed', [])
+                                    self._indicator_results['not_saved'] = self._indicator_results.get(
+                                        'failed', [])
                                     self._indicator_results['saved'].extend(
                                         [i.get('summary') for i in chunk
                                          if i.get('summary') not in self._indicator_results.get(
@@ -599,7 +600,7 @@ class TcExJob(object):
                     resource.indicator(i_data.get('summary'))
                 else:
                     resource.indicator(i_value)
-                tag_resource = resource.tags(self._tcex.safetag(tag.get('name')))
+                tag_resource = resource.tags(tag.get('name'))
                 tag_resource.http_method = 'POST'
                 t_results = tag_resource.request()
                 if t_results.get('status') != 'Success':
@@ -1117,7 +1118,7 @@ class TcExJob(object):
             del self._group_cache[owner][resource_type][name]
 
     def group_cache_type(self, group_id, owner):
-        """Get the group type for the provided group id
+        """Get the group type for the provided group id.
 
         **Cache Structure**
         ::
@@ -1163,7 +1164,7 @@ class TcExJob(object):
 
     @property
     def group_results(self):
-        """Result dictionary of failed, saved, not_saved, and submitted groups
+        """Result dictionary of failed, saved, not_saved, and submitted groups.
 
         Returns:
             (dictionary): Dictionary of group names for each status
@@ -1225,7 +1226,7 @@ class TcExJob(object):
 
     @property
     def unprocessed_indicators(self):
-        """ Return indicators (unprocessed)
+        """Return indicators (unprocessed).
 
         Returns:
             (list): The unprocessed indicator list
@@ -1234,7 +1235,7 @@ class TcExJob(object):
 
     @unprocessed_indicators.setter
     def unprocessed_indicators(self, value):
-        """ Set indicators list
+        """Set indicators list.
 
         Args:
             value (list): List of indicators
@@ -1253,7 +1254,7 @@ class TcExJob(object):
 
     @property
     def indicator_len(self):
-        """The current length of the indicator list
+        """The current length of the indicator list.
 
         Returns:
             (integer): The length of the indicator list
@@ -1262,7 +1263,7 @@ class TcExJob(object):
 
     @property
     def indicator_results(self):
-        """Result dictionary of failed, saved, not_saved, and submitted indicators
+        """Result dictionary of failed, saved, not_saved, and submitted indicators.
 
         Returns:
             (dictionary): Dictionary of indicator values for each status
@@ -1270,9 +1271,12 @@ class TcExJob(object):
         return self._indicator_results
 
     def process(self, owner, indicator_batch=True, group_action='skip'):
-        """Process all association, group and indicator data.
+        """Process all groups, indicator data, and associations.
 
-        Process each of the supported data types for this job.
+        Process each of the supported data types for this job, in the following
+        order (left to right):
+
+            groups > indicators > file occurrences > group associations > associations
 
         Args:
             owner (string): The owner name for the data to be written