+++ tcex/tcex_utils.py
@@ -160,10 +160,10 @@ class TcExUtils():
             src_tzname = source_datetime.tzname()
         dt, status = cal.parseDT(time_input, sourceTime=source_datetime, tzinfo=tzinfo)
         if tz is not None:  # don't add tz if no tz value is passed
-            # don't covert timezone if source timezone already in the correct timezone
             if dt.tzinfo is None:
                 self._tcex.log.info('Assuming local time for naive datetime {}.'.format(str(dt)))
                 dt = dt.replace(tzinfo=timezone(get_localzone().zone))  # required for py2.x
+            # don't covert timezone if source timezone already in the correct timezone
             if tz != src_tzname:
                 dt = dt.astimezone(timezone(tz))
         if status == 0: